code/requirements.txt
opencv-python-headless==4.11.0.86
Pillow==11.0.0
Cython==3.1.2
pycocotools==2.0.10
datasets==4.0.0
huggingface_hub==0.33.5
tensorboard==2.20.0
matplotlib==3.10.3
plotnine==0.15.0
seaborn==0.12.0
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.6
tqdm==4.66.3
pyyaml==6.0.1
ujson==5.10.0
requests==2.32.2
faster-coco-eval==1.6.7
onnx==1.18.0
onnxruntime==1.22.1
scikit-learn==1.7.1
numpy==1.26.4
torch==2.7.1
torchvision==0.22.1
--extra-index-url https://download.pytorch.org/whl/cu126
